Fix missing values when parsing App Manifest
For some reason, retrieving AndroidManifest.xml from first element
of AssetManager.openXmlResourceParser() won't always return the
expected data. This prevents shortcuts and/or round icons to be
parsed for some apps (e.g. News, Maps, Photos, Youtube). Make a
reflection call to addAssetPath() so that the proper cookie is used.